// Dunyana.Web/src/app/pages/customer/model/DTOs/WalletDto.ts
export class WalletDto {
    id: number;
    customerID: string;
    type: number;
    transaction: string;
    transactionDate: string;
    transactionAmount: number;
    transactionDescription: string;
    status: number;
    orderID: number;
    saveCard: boolean;
    sourceId: any;
    paymentResponseSource: any;
    isExistingCard: boolean;
    orderWalletPaymentAmount: any;


}